Reorg packages into core, jobmon, cluster
