--- docs/source/api_reference/cli_tools.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium_cluster_tools.cli_tools
