+++ docs/source/api_reference/core/cli_tools.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium_cluster_tools.core.cli_tools
