+++ docs/source/api_reference/core/cluster/cli_options.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium_cluster_tools.core.cluster.cli_options
