+++ docs/source/api_reference/core/cluster/index.rst
@@ -1,0 +1,7 @@
+.. automodule:: vivarium_cluster_tools.core.cluster
+
+.. toctree::
+   :maxdepth: 2
+   :glob:
+
+   *
