+++ docs/source/api_reference/core/cluster/interface.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium_cluster_tools.core.cluster.interface
