+++ docs/source/api_reference/core/index.rst
@@ -1,0 +1,8 @@
+.. automodule:: vivarium_cluster_tools.core
+
+.. toctree::
+   :maxdepth: 2
+   :glob:
+
+   *
+   */index
