--- docs/source/api_reference/logs.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium_cluster_tools.logs
