--- docs/source/api_reference/psimulate/cluster/cli_options.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium_cluster_tools.psimulate.cluster.cli_options
