--- docs/source/api_reference/psimulate/cluster/index.rst
@@ -1,7 +1,0 @@
-.. automodule:: vivarium_cluster_tools.psimulate.cluster
-
-.. toctree::
-   :maxdepth: 2
-   :glob:
-
-   *
