--- docs/source/api_reference/psimulate/cluster/interface.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium_cluster_tools.psimulate.cluster.interface
