+++ src/vivarium_cluster_tools/core/__init__.py
@@ -1,0 +1,10 @@
+"""
+====
+core
+====
+
+Shared infrastructure consumed by both the ``psimulate`` and ``dagger`` CLIs:
+cluster validation and resource specifications, the Jobmon SDK facade,
+Slack notifications, shared Click decorators, and logging configuration.
+
+"""
