+++ src/vivarium_cluster_tools/core/jobmon/__init__.py
@@ -1,0 +1,11 @@
+"""
+======
+jobmon
+======
+
+Jobmon integration layer: SDK facade (:mod:`.client`), env resolution
+(:mod:`.env`), and shared Click decorators (:mod:`.cli_options`).
+
+"""
+
+from vivarium_cluster_tools.core.jobmon.cli_options import with_max_attempts, with_max_workers
