+++ src/vivarium_cluster_tools/dagger/workflow_config/utilities.py
@@ -17,9 +17,9 @@ from datetime import datetime
 from pathlib import Path
 from typing import TYPE_CHECKING
 
+from vivarium_cluster_tools.core.jobmon import client
+from vivarium_cluster_tools.core.jobmon.env import resolve_env_prefix
 from vivarium_cluster_tools.dagger.workflow_config.config import ResourceConfig
-from vivarium_cluster_tools.psimulate.jobmon_config import client
-from vivarium_cluster_tools.psimulate.jobmon_config.workflow import resolve_env_prefix
 
 if TYPE_CHECKING:
     from jobmon.client.api import Tool
