--- src/vivarium_cluster_tools/psimulate/jobmon_config/__init__.py
@@ -1,13 +1,0 @@
-"""
-========================
-Jobmon Configuration
-========================
-
-Configuration and utilities for Jobmon-based job scheduling.
-
-"""
-
-from vivarium_cluster_tools.psimulate.jobmon_config.cli_options import (
-    with_max_attempts,
-    with_max_workers,
-)
