+++ src/vivarium_cluster_tools/psimulate/runner.py
@@ -20,19 +20,18 @@ import yaml
 from loguru import logger
 from vivarium.framework.utilities import collapse_nested_dict
 
-from vivarium_cluster_tools import logs
-from vivarium_cluster_tools.notifications import send_slack_notification
+from vivarium_cluster_tools.core import cluster, logs
+from vivarium_cluster_tools.core.jobmon import client
+from vivarium_cluster_tools.core.notifications import send_slack_notification
 from vivarium_cluster_tools.psimulate import (
     COMMANDS,
     branches,
-    cluster,
     jobs,
     model_specification,
     paths,
     pip_env,
 )
-from vivarium_cluster_tools.psimulate.jobmon_config import client
-from vivarium_cluster_tools.psimulate.jobmon_config.workflow import build_workflow
+from vivarium_cluster_tools.psimulate.jobmon_workflow import build_workflow
 from vivarium_cluster_tools.psimulate.paths import OutputPaths
 from vivarium_cluster_tools.psimulate.performance_logger import (
     append_perf_data_to_central_logs,
