+++ tests/psimulate/cluster/test_interface.py
@@ -6,13 +6,13 @@ from unittest.mock import patch
 
 import pytest
 
-from vivarium_cluster_tools.psimulate.cluster.interface import (
+from vivarium_cluster_tools.core.cluster.interface import (
     _SLURM_TIMEOUT_BUFFER_SECONDS,
     NativeSpecification,
     _parse_slurm_time,
     get_workflow_timeout_seconds,
 )
-from vivarium_cluster_tools.psimulate.cluster.validation import validate_cluster_environment
+from vivarium_cluster_tools.core.cluster.validation import validate_cluster_environment
 
 
 @pytest.fixture(
@@ -201,7 +201,7 @@ class TestGetRunnerNodeRemainingSeconds:
         """Return remaining seconds minus the safety buffer."""
         completed = _make_squeue_result("10:00:00")
         with patch(
-            "vivarium_cluster_tools.psimulate.cluster.interface.subprocess.run",
+            "vivarium_cluster_tools.core.cluster.interface.subprocess.run",
             return_value=completed,
         ):
             result = get_workflow_timeout_seconds()
@@ -211,7 +211,7 @@ class TestGetRunnerNodeRemainingSeconds:
         """Raise when SLURM reports UNLIMITED time."""
         completed = _make_squeue_result("UNLIMITED")
         with patch(
-            "vivarium_cluster_tools.psimulate.cluster.interface.subprocess.run",
+            "vivarium_cluster_tools.core.cluster.interface.subprocess.run",
             return_value=completed,
         ):
             with pytest.raises(ValueError, match="Unrecognized SLURM time format"):
@@ -226,7 +226,7 @@ class TestGetRunnerNodeRemainingSeconds:
         """Raise when squeue returns a non-numeric time string."""
         completed = _make_squeue_result(bad_value)
         with patch(
-            "vivarium_cluster_tools.psimulate.cluster.interface.subprocess.run",
+            "vivarium_cluster_tools.core.cluster.interface.subprocess.run",
             return_value=completed,
         ):
             with pytest.raises(ValueError, match="Unrecognized SLURM time format"):
@@ -237,7 +237,7 @@ class TestGetRunnerNodeRemainingSeconds:
         remaining_seconds = _SLURM_TIMEOUT_BUFFER_SECONDS - 1
         completed = _make_squeue_result(self.convert_seconds_to_time_str(remaining_seconds))
         with patch(
-            "vivarium_cluster_tools.psimulate.cluster.interface.subprocess.run",
+            "vivarium_cluster_tools.core.cluster.interface.subprocess.run",
             return_value=completed,
         ):
             with pytest.raises(RuntimeError, match="Not enough time"):
@@ -250,7 +250,7 @@ class TestGetRunnerNodeRemainingSeconds:
             self.convert_seconds_to_time_str(_SLURM_TIMEOUT_BUFFER_SECONDS)
         )
         with patch(
-            "vivarium_cluster_tools.psimulate.cluster.interface.subprocess.run",
+            "vivarium_cluster_tools.core.cluster.interface.subprocess.run",
             return_value=completed,
         ):
             with pytest.raises(RuntimeError, match="Not enough time"):
@@ -260,7 +260,7 @@ class TestGetRunnerNodeRemainingSeconds:
         remaining_seconds = _SLURM_TIMEOUT_BUFFER_SECONDS + 1
         completed = _make_squeue_result(self.convert_seconds_to_time_str(remaining_seconds))
         with patch(
-            "vivarium_cluster_tools.psimulate.cluster.interface.subprocess.run",
+            "vivarium_cluster_tools.core.cluster.interface.subprocess.run",
             return_value=completed,
         ):
             assert get_workflow_timeout_seconds() == 1
@@ -268,7 +268,7 @@ class TestGetRunnerNodeRemainingSeconds:
     def test_raises_on_subprocess_error(self) -> None:
         """Raise when squeue fails."""
         with patch(
-            "vivarium_cluster_tools.psimulate.cluster.interface.subprocess.run",
+            "vivarium_cluster_tools.core.cluster.interface.subprocess.run",
             side_effect=FileNotFoundError("squeue not found"),
         ):
             with pytest.raises(RuntimeError, match="Could not determine"):
@@ -292,7 +292,7 @@ class TestGetRunnerNodeRemainingSeconds:
             stderr="slurm_load_jobs error: Invalid job id specified\n",
         )
         with patch(
-            "vivarium_cluster_tools.psimulate.cluster.interface.subprocess.run",
+            "vivarium_cluster_tools.core.cluster.interface.subprocess.run",
             return_value=completed,
         ):
             with pytest.raises(RuntimeError, match="squeue failed"):
@@ -302,7 +302,7 @@ class TestGetRunnerNodeRemainingSeconds:
         """Raise when squeue returns empty output."""
         completed = _make_squeue_result("")
         with patch(
-            "vivarium_cluster_tools.psimulate.cluster.interface.subprocess.run",
+            "vivarium_cluster_tools.core.cluster.interface.subprocess.run",
             return_value=completed,
         ):
             with pytest.raises(RuntimeError, match="no output"):
@@ -312,7 +312,7 @@ class TestGetRunnerNodeRemainingSeconds:
         """Handle D-HH:MM:SS format from squeue."""
         completed = _make_squeue_result("1-12:00:00")
         with patch(
-            "vivarium_cluster_tools.psimulate.cluster.interface.subprocess.run",
+            "vivarium_cluster_tools.core.cluster.interface.subprocess.run",
             return_value=completed,
         ):
             result = get_workflow_timeout_seconds()
